plantback/requirements.txt
# python 3.11
asyncio==3.4.3
blinker==1.9.0
click==8.1.7
DBUtils==3.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paho-mqtt==2.1.0
PyJWT==2.10.0
PyMySQL==1.1.1
tornado==6.4.2
websocket-server==0.6.4
websockets==14.1
Werkzeug==3.1.3
